Fix path to AO2D for AnalysisQC

+++ DATA/production/configurations/2021/OCT/apass4/async_pass.sh
@@ -170,9 +170,13 @@ if [[ -f "AO2D.root" ]]; then
 	echo "exit code from AO2D check is " $exitcode
 	exit $exitcode
     fi
-    ${O2DPG_ROOT}/MC/analysis_testing/o2dpg_analysis_test_workflow.py --merged-task -f ../../AO2D.root
+    ${O2DPG_ROOT}/MC/analysis_testing/o2dpg_analysis_test_workflow.py --merged-task -f AO2D.root
     ${O2DPG_ROOT}/MC/bin/o2_dpg_workflow_runner.py -f workflow_analysis_test.json
-    mv Analysis/MergedAnalyses/AnalysisResults.root .
+    if [[ -f "Analysis/MergedAnalyses/AnalysisResults.root" ]]; then
+       mv Analysis/MergedAnalyses/AnalysisResults.root .
+    else
+	echo "No Analysis/MergedAnalyses/AnalysisResults.root found! check analysis QC"
+    fi
 fi
 
 # copying the QC json file here
